Fix otpVerification Page:Added PlaceHolder and moved it from component to pages folder
